refactored CornGrowEvent and CornHarvestEvent into Tasks. added todo
